[WIP] Allow ntex to use external buffer pools.
(This is still a WIP change and not ready to be merged. I'm opening this PR
to discuss the changes to ntex-bytes and ntex-codec while I work on ntex and
ntex-mqtt.)

ntex-bytes now has two new traits, BufferSource and BufferSink.
A BufferSource is a way to asynchronously receive a buffer of a minimum
requested size. Such a buffer internally holds on to a reference to
a BufferSink, such that when the buffer is Drop'd it'll automatically return
itself to the sink. The source and sink will normally be the same type; they're
split up into two different traits because some parts of the stack only care
about sources (eg encoders) and others only about sinks (eg decoders and
frames). There is a trait alias BufferPool for BufferSource + BufferSink
for convenience.

A BufferPool deals with two types of buffers, owned and shared buffers.
An owned buffer implements the Owned trait and is generally equivalent to
bytes::BytesMut - it is the sole owner of a range of contiguous bytes and thus
provides mutable access to those bytes, along with a read and write cursor
to mark what subset of the range contains readable bytes. An owned buffer
can be split into smaller owned buffers with split_to, and can be frozen into
a shared buffer with freeze. A shared buffer implements the Shared trait and
is generally equivalent to bytes::Bytes - it is not necessarily the sole owner
of the range of bytes and thus can only provide readonly access.

Unlike BytesMut, an Owned has no requirement of being able to grow to fit more
bytes in it than it has the capacity for. Also, by allowing the types to be
externally pluggable, it allows users like us to use more optimized types for
our use case, such as types that do not have an internal vtable to switch
between Vec/Arc backing implementations, especially when such switching
involves memcpy'ing.

Lastly, a buffer pool has a third associated type called an accumulator that
implements the BytesAccumulator trait. This trait is intended to be used by
Encoders and is a sink for Shared buffers. Unlike a BytesMut that can only
merge Bytes into itself by copying them, a BytesAccumulator implementation
might choose to maintain a Vec<Shared>-like rope structure that can be used
with vectored writes.

Because the pool needs to be threaded through codecs and servers, many types
now have an additional type parameter P for the pool type. As Nikolay had
requested, there is a default pool whose Owned and Accumulator are BytesMut and
whose Shared is Bytes - ntex_bytes::BytesPool - and the type parameter
defaults to this pool. However, there are still API changes to account for
the fact that receiving a buffer from the pool is asynchronous.

- ntex_codec::Encoder::encode now takes a pool and an accumulator instead of
  a BytesMut, and returns a Future of the encoding result instead of the result
  directly.

- As a consequence ntex_codec::Framed::encode is also asynchronous.

- ntex_codec::Framed::encode now supports using vectored write when encoding, if
  the IO type and the BytesAccumulator support it.

- ntex_codec::Framed::next_item is now asynchronous.

- ntex_codec::Framed no longer impl Stream and Sink. I've deleted these impls
  for now. The alternative would be to add newtype wrappers that hold on to
  the intermediate futures from ::encode and ::next_item.

Other changes:

- Dependencies on other ntex- crates in the same repo now use both
  `version = "..."` and `path = "../..."`. This allows testing local changes.