feat: allow extracting TLS SNI server name and TLS PSK identity
This change allows to query for the SNI server name and the PSK
identity.

The server name implements only the DNS variant, as this is what both
APIs currently seem to support.

The PSK identity is only implemented for OpenSSL, as RustTLS currently
doesn't support TLS PSK. But it could be implemented the same way, once
this becomes available.